requirements-dev.txt
black
mypy
pytest
pytest-mock
pytest-django
pytest-cov
pytest-asyncio